Use slim distribution for Java client downloads
There were some comments about the download size of the Java client
in https://hazelcast.com/clients/java/.

This was due to the fact that we were using the normal distribution for
the download link, starting from 5.0.

For the client, we believe the slim distribution as the download link
should be enough. This PR changes the already added links with the
slim distribution.

Ondrej will fix the release automation so that future additions
to this file will use the slim distribution for clients.