Test URLs are valid at PR time
The post-release checklist includes a manual check for the download links - validated by spotting [an error](
https://github.com/hazelcast/rel-scripts/commit/013b2d54ec3076221e6f352d4d9228e1c0047ff8#diff-0ae958136aed79d39220a53fc0899485e646ddb3f63a524771a41669de1c4686R14) in a URL.

These checks can be automated

Note - this is triggered by a `pull_request`, but the release pipeline `push`es directly. Although we could trigger based on `push`, it would be unhelpful to notify a bot GitHub Actor of a failure, so this is an _additional_ check alongside the manual one.