Test URLs are valid at PR time [DI-334] (#33)
The post-release checklist includes a manual check for the download
links - validated by spotting [an error](
https://github.com/hazelcast/rel-scripts/commit/013b2d54ec3076221e6f352d4d9228e1c0047ff8#diff-0ae958136aed79d39220a53fc0899485e646ddb3f63a524771a41669de1c4686R14)
in a URL.

These checks can instead be automated.

Added an action that, on a PR, tests the updated _lines_ for URL-alike
content, and if found, tests they actually resolve - flagging those that
don't.
It doesn't test _all_ the content because some links don't work on
purpose - e.g. the
[5.4.0-BETA-2](https://github.com/hazelcast/rel-scripts/blob/49b2bc718df3f1a8c5d8095e168728a21a1e4101/hazelcast-open-source.txt#L531C19-L531C116)
URL is listed in a (now hidden) section, but the content has since been
deleted. But we can assume when _adding_ new content, the links should
at least work at that point.

This is triggered by a `pull_request`, but the release pipeline `push`es
directly. Although we _could_ trigger based on `push`, it would be
unhelpful to notify a bot GitHub Actor of a failure, so this is an
_additional_ check alongside the manual one.

Execution examples:
-
[Success](https://github.com/JackPGreen/rel-scripts/actions/runs/11712124649/job/32622121495)
-
[Failure](https://github.com/JackPGreen/rel-scripts/actions/runs/11711877446)
(with error from
https://github.com/hazelcast/rel-scripts/commit/013b2d54ec3076221e6f352d4d9228e1c0047ff8)

Fixes: [DI-334](https://hazelcast.atlassian.net/browse/DI-334)

[DI-334]:
https://hazelcast.atlassian.net/browse/DI-334?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ